% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprFuncs.R
\name{cpmGeneExpr}
\alias{cpmGeneExpr}
\title{Calculate CPM values for a tibble of counts and return specific genes only}
\usage{
cpmGeneExpr(counts, gene.col, goi, log = TRUE)
}
\arguments{
\item{counts}{Counts as tibble (gene or transcript ID in a column)}

\item{gene.col}{The name of the column that has the gene names or transcript IDs}

\item{goi}{Genes interested in finding CPM values for}

\item{log}{Whether you want logCPM or not. Default is TRUE}
}
\value{
A tibble with CPM for genes of interest
}
\description{
Calculate CPM values for a tibble of counts and return specific genes only
}
\examples{
cpmGeneExpr(counts, gene.col = "Genes", goi = c("Gene_1","Gene_2"))
}
