% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_to_genes.R
\name{ensemblToGenes}
\alias{ensemblToGenes}
\title{Takes a dataframe and a column that has Ensembl IDs (or rownames) and returns dataframe with a column of Gene IDs.}
\usage{
ensemblToGenes(data, ...)
}
\arguments{
\item{data}{A dataframe}

\item{column}{The name of the column that has the Ensembl IDs. If IDs are stored as rownames, use column = "rownames"}

\item{type}{Whether the type of Ensembl IDs are transcript or gene}

\item{keep.ensembl.ids}{Whether to keep Ensembl IDs as a separate column}

\item{make.genes.unique}{Choose whether you want only unique gene names. Duplicated genes will be separated by a dash. If IDs are stored as rownames,
make.genes.unique must be TRUE}
}
\value{
A dataframe with column now as Ensembl IDs
}
\description{
Takes a dataframe and a column that has Ensembl IDs (or rownames) and returns dataframe with a column of Gene IDs.
}
\examples{
ensemblToGenes(data = counts.table, column = "Genes", type = "gene")
}
